stratifier: fix fractional difficulty support for sub-1 startdiff/mindiff
client->diff, old_diff, and suggest_diff were typed int64_t, truncating
fractional values (e.g. 0.001 → 0) at client initialization. The pool
was sending mining.set_difficulty(0) to miners, which they ignored in
favour of their firmware default (~1M).

Changes:
- stratifier.c: change client diff/old_diff/suggest_diff from int64_t
  to double
- stratifier.c: remove (int64_t) cast on ckp->startdiff at client init
- stratifier.c: fix stratum_send_diff JSON pack format I→f so the wire
  value is sent as double, not truncated int64
- stratifier.c: fix add_submit vardiff — change optimal to double,
  remove lround() and all (int64_t) casts on mindiff/maxdiff, lower
  the sub-1 guard from < 1 to < ckp->mindiff, fix log format PRId64→%g
- stratifier.c: fix suggest_diff() — local sdiff int64_t→double,
  sscanf PRId64→%lf, accept both integer and real JSON params,
  remove (int64_t) cast on ckp->mindiff clamp
- stratifier.c: fix worker mindiff on auth — remove (int64_t) casts,
  extend guard from >= 1.0 to > 0 to allow sub-1 worker mindiffs
- stratifier.c: fix proxy update_diff guard from < 1 to < 0.0001
- ckpool.h: change highdiff from int64_t to double for type consistency
- ckpool.c: use json_get_num_as_double for highdiff to match new type
- generator.c: fix suggest_diff JSON pack format I→f, remove cast

+++ src/ckpool.h
@@ -235,7 +235,7 @@ struct ckpool_instance {
 	/* Difficulty settings */
 	double mindiff; // Default 1
 	double startdiff; // Default 42
-	int64_t highdiff; // Default 1000000
+	double highdiff; // Default 1000000
 	double maxdiff; // No default (0 = unlimited)
 
 	/* Coinbase data */
